moved event utilities to events.go
